app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "python-dotenv>=1.0.0",
]
